drivers: mpsl: Implement onoff API for MPSL
This implements the onoff interface for the MPSL clock control driver.
The MPSL driver will still do reference counting of API usage across
calls via the onoff interface and the normal clock control driver
interface. It will *not* return an error if, e.g., the onoff_start is
requested twice in a row, or if the onoff interface and a user of the
driver interface requests different operations.

In practice, there are no guarantees that what the onoff interface
requests from the driver will be the end result. This is because the
MPSL clock driver may internally be requested to run, even though there
are no outside users.

Ref.: DRGN-14661

Signed-off-by: Thomas Stenersen <thomas.stenersen@nordicsemi.no>